Create image_editor.cpp
 Given an image you will ﬁrst read the image into an array, perform some basic operations and generate the output image.
Note: the input file must be  provided in pgm (portable graymap format).  These image ﬁles have the following speciﬁcations. When you open images in pgm format using any text editor, you will see the following. The ﬁrst line contains an identiﬁer for the image which for this assignment will be P2. The next line contains dimension of the image with two numbers indicating the width and the height respectively. Next number speciﬁes the maximum allowed pixel value of the image, for this assignment it will be 255. Next we have the actual pixel values of the image given in row major ordering. That is the ﬁrst pixel value corresponds to location (0,0), the second for location (0,1) and so on. For all pixels in the image, we have their value between 0 and 255. The number of pixel values equals width times height of the image.